+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,0 +1,5 @@
+package calculator
+
+fun interface Calculator {
+    fun calculate(expression: Expression): Double
+}
