+++ src/test/kotlin/calculator/StringCalculatorTest.kt
@@ -23,18 +23,3 @@ class StringCalculatorTest : StringSpec({
         }
     }
 })
-
-interface Calculator {
-    fun calculate(expression: Expression): Double
-}
-
-class StringCalculator : Calculator {
-    override fun calculate(expression: Expression): Double {
-        val numbers = expression.extractNumbers()
-        val operators = expression.extractOperators()
-
-        return operators.foldIndexed(numbers.first()) { index, acc, operator ->
-            operator.apply(acc, numbers[index + 1])
-        }
-    }
-}
